made improvements

+++ src/main/java/com/diabetesapp/model/DiabetesRecord.java
@@ -14,11 +14,11 @@ public class DiabetesRecord {
 
     private int diabetesRecordId;
 
-    private int bloodGlucoseLevel;
+    private double bloodGlucoseLevel;
 
-    private int carbIntake;
+    private double carbIntake;
 
-    private int medicationDose;
+    private double medicationDose;
 
     private Date dateRecorded;
